Expanded authentication sample
