+++ sdk/textanalytics/ai-text-analytics/samples/javascript/authenticationMethods.js
@@ -1,0 +1,67 @@
+// Copyright (c) Microsoft Corporation.
+// Licensed under the MIT License.
+
+/**
+ * Demonstrates how to use either an Azure Active Directory (RBAC) or an
+ * API Key to authenticate a TextAnalyticsClient
+ */
+
+// To use an API Key, import `AzureKeyCredential` from the Text Analytics package
+const { TextAnalyticsClient, AzureKeyCredential } = require("@azure/ai-text-analytics");
+
+// To use Azure AD, import `DefaultAzureCredential` from the `@azure/identity` package
+const { DefaultAzureCredential } = require("@azure/identity");
+
+// Load the .env file if it exists
+const dotenv = require("dotenv");
+dotenv.config();
+
+// You will need to set this environment variables or edit the following values
+const endpoint = process.env["ENDPOINT"] || "<cognitive services endpoint>";
+
+async function useAad() {
+  console.log("-- Azure Active Directory --");
+
+  // DefaultAzureCredential expects the following three environment variables:
+  // - AZURE_TENANT_ID: The tenant ID in Azure Active Directory
+  // - AZURE_CLIENT_ID: The application (client) ID registered in the AAD tenant
+  // - AZURE_CLIENT_SECRET: The client secret for the registered application
+  const credential = new DefaultAzureCredential();
+
+  const client = new TextAnalyticsClient(endpoint, credential);
+
+  const [result] = await client.detectLanguage(["hello world"]);
+
+  if (!result.error) {
+    console.log(`Primary language detected as ${result.primaryLanguage.name}`);
+  }
+}
+
+async function useApiKey() {
+  console.log("-- API Key --");
+
+  // If using an API Key, you will need to set this environment variable
+  const apiKey = process.env["TEXT_ANALYTICS_API_KEY"] || "<api key>";
+
+  const credential = new AzureKeyCredential(apiKey);
+
+  const client = new TextAnalyticsClient(endpoint, credential);
+
+  const [result] = await client.detectLanguage(["hello world"]);
+
+  if (!result.error) {
+    console.log(`Primary language detected as ${result.primaryLanguage.name}`);
+  }
+}
+
+async function main() {
+  console.log("== Client Authentication Methods Sample ==");
+
+  await useAad();
+
+  await useApiKey();
+}
+
+main().catch((err) => {
+  console.error("The sample encountered an error:", err);
+});
